Effect_Att/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=2.1.0=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
anyio=4.3.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
babel=2.14.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blas=2.116=mkl
blas-devel=3.9.0=16_linux64_mkl
bleach=6.1.0=pypi_0
bob.blitz=2.0.24=py38h5652565_0
bob.core=2.2.8=py38h5652565_0
bob.extension=6.1.1=py38hfe1193c_0
bob.io.base=4.0.0=py38h33f8406_0
bob.learn.activation=2.0.18=py38h5652565_0
bob.learn.linear=2.1.10=py38hd1e3326_0
bob.math=3.1.9=py38h6772507_0
boost=1.76.0=py38h2b96118_1
boost-cpp=1.76.0=h312852a_1
brotli-python=1.1.0=py38h17151c0_1
bzip2=1.0.8=hd590300_5
c-ares=1.27.0=hd590300_0
ca-certificates=2024.2.2=hbcca054_0
cachetools=4.2.4=pypi_0
certifi=2024.2.2=pyhd8ed1ab_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pyhd8ed1ab_0
click=8.1.7=unix_pyh707e725_0
click-plugins=1.1.1=py_0
comm=0.2.2=pypi_0
cuda-cudart=12.1.105=0
cuda-cudart_linux-64=12.4.99=h59595ed_0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.99=0
cuda-runtime=12.1.0=0
cuda-version=12.4=h3060b56_3
cudnn=8.9.7.29=h092f7fd_3
cycler=0.12.1=pypi_0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pypi_0
fastjsonschema=2.19.1=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=pyhd8ed1ab_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=h267a509_2
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pyhca7485f_0
gitdb=4.0.11=pypi_0
gitpython=3.1.2=pypi_0
gmp=6.3.0=h59595ed_1
gmpy2=2.1.2=py38h793c122_1
gnutls=3.6.13=h85f3911_1
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.62.1=pypi_0
h11=0.14.0=pypi_0
hdf5=1.10.6=h3ffc7dd_1
httpcore=1.0.4=pypi_0
httpx=0.27.0=pypi_0
icu=68.2=h9c3ff4c_0
idna=3.6=pyhd8ed1ab_0
importlib-metadata=7.1.0=pypi_0
importlib-resources=6.4.0=pypi_0
ipykernel=6.29.3=pypi_0
ipython=8.12.3=pypi_0
ipywidgets=8.1.2=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.3=pyhd8ed1ab_0
joblib=1.3.2=pypi_0
jpeg=9e=h166bdaf_2
json5=0.9.24=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.6.1=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.4=pypi_0
jupyter-server=2.13.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.1.5=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.25.4=pypi_0
jupyterlab-widgets=3.0.10=pypi_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.5=pypi_0
krb5=1.21.2=h659d440_0
lame=3.100=h166bdaf_1003
lcms2=2.15=hfd0df8a_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libabseil=20230802.1=cxx17_h59595ed_0
libblas=3.9.0=16_linux64_mkl
libblitz=1.0.2=h9c3ff4c_3
libcblas=3.9.0=16_linux64_mkl
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.9.0.20=0
libcurand=10.3.5.119=hd3aeb46_0
libcurl=8.6.0=hca28451_0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h0b41bf4_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=hd590300_2
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h807b86a_5
libgfortran-ng=13.2.0=h69a702a_5
libgfortran5=13.2.0=ha4646dd_5
libiconv=1.17=hd590300_2
libjpeg-turbo=2.0.0=h9bf148f_0
liblapack=3.9.0=16_linux64_mkl
liblapacke=3.9.0=16_linux64_mkl
libmagma=2.7.2=h173bb3b_2
libmagma_sparse=2.7.2=h173bb3b_3
libnghttp2=1.58.0=h47da74e_1
libnpp=12.0.2.50=0
libnsl=2.0.1=hd590300_0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libopenblas=0.3.25=pthreads_h413a1c8_0
libpng=1.6.43=h2797004_0
libprotobuf=4.25.1=hf27288f_2
libsqlite=3.45.2=h2797004_0
libssh2=1.11.0=h0841786_0
libstdcxx-ng=13.2.0=h7e041cc_5
libtiff=4.5.0=h6adf6a1_2
libuuid=2.38.1=h0b41bf4_0
libuv=1.48.0=hd590300_0
libwebp-base=1.3.2=hd590300_0
libxcb=1.13=h7f98852_1004
libxcrypt=4.4.36=hd590300_1
libzlib=1.2.13=hd590300_5
llvm-openmp=15.0.7=h0cdce71_0
magma=2.7.2=h51420fd_3
markdown=3.6=pypi_0
markupsafe=2.1.5=py38h01eb140_0
matplotlib=3.3.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=3.0.2=pypi_0
mkl=2022.1.0=h84fe81f_915
mkl-devel=2022.1.0=ha770c72_916
mkl-include=2022.1.0=h84fe81f_915
mpc=1.3.1=hfe3b2da_0
mpfr=4.2.1=h9458935_0
mpmath=1.3.0=pyhd8ed1ab_0
multidict=6.0.5=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.3=pypi_0
nbformat=5.10.3=pypi_0
nccl=2.20.5.1=h3a97aeb_0
ncurses=6.4.20240210=h59595ed_0
nest-asyncio=1.6.0=pypi_0
nettle=3.6=he412f7d_0
networkx=2.5=pypi_0
ninja=1.11.1=h924138e_0
notebook=7.1.2=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.24.4=py38h59b608b_0
oauthlib=3.2.2=pypi_0
openh264=2.1.1=h780b84a_0
openjpeg=2.5.0=hfec8fc6_2
openssl=3.2.1=hd590300_1
overrides=7.7.0=pypi_0
packaging=24.0=pyhd8ed1ab_0
pandocfilters=1.5.1=pypi_0
parso=0.8.3=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=py38hde6dc18_1
pip=24.0=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=4.2.0=pyhd8ed1ab_0
pooch=1.8.1=pyhd8ed1ab_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=3.20.0=pypi_0
psutil=5.9.8=pypi_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycairo=1.20.0=pypi_0
pycparser=2.21=pypi_0
pygments=2.17.2=pypi_0
pykalman=0.9.7=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.8.19=hd12c33a_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python-igraph=0.8.3=pypi_0
python-json-logger=2.0.7=pypi_0
python_abi=3.8=4_cp38
pytorch=2.2.1=py3.8_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=py38h01eb140_1
pyzmq=25.1.2=pypi_0
qtconsole=5.5.1=pypi_0
qtpy=2.4.1=pypi_0
readline=8.2=h8228510_1
referencing=0.34.0=pypi_0
requests=2.31.0=pyhd8ed1ab_0
requests-oauthlib=2.0.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.18.0=pypi_0
rsa=4.9=pypi_0
scikit-learn=0.24.0=pypi_0
scipy=1.5.4=pypi_0
send2trash=1.8.2=pypi_0
setuptools=69.2.0=pyhd8ed1ab_0
six=1.16.0=pypi_0
sleef=3.5.1=h9b69904_2
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
stack-data=0.6.3=pypi_0
sympy=1.12=pypyh9d50eac_103
tbb=2021.7.0=h924138e_0
tensorboard=2.2.2=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
terminado=0.18.1=pypi_0
texttable=1.7.0=pypi_0
threadpoolctl=3.4.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.13=noxft_h4845f30_101
tomli=2.0.1=pypi_0
torch-geometric=2.5.2=pypi_0
torchaudio=2.2.1=py38_cu121
torchtriton=2.2.0=py38
torchvision=0.17.1=py38_cu121
tornado=6.4=pypi_0
tqdm=4.66.2=pypi_0
traitlets=5.14.2=pypi_0
types-python-dateutil=2.9.0.20240316=pypi_0
typing_extensions=4.10.0=pyha770c72_0
uri-template=1.3.0=pypi_0
urllib3=2.2.1=pyhd8ed1ab_0
wcwidth=0.2.13=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.7.0=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.43.0=pyhd8ed1ab_0
widgetsnbextension=4.0.10=pypi_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.6=h166bdaf_0
yaml=0.2.5=h7f98852_2
yarl=1.9.4=pypi_0
zipp=3.18.1=pypi_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hfc55251_0
